Funzt lola

+++ src/main/java/data/InputInfos.java
@@ -3,5 +3,5 @@ package data;
 public class InputInfos {
     public Integer[] mousePos = {0,0};
     public Boolean[] mouseClick = {false,false};
-    public Integer mouseWheel = 0;
+    public Double mouseWheel = 0d;
 }
